+++ src/ScreenHook.cpp
@@ -317,7 +317,7 @@ void TextHook::Draw(void) {
     if (GetIsVisible() && GetX() != -1 && GetY() != -1 && text) {
         uint x = GetX(), y = GetY(), w = CalculateTextLen(text, font).x;
         x -= (alignment != Center ? (alignment != Right ? 0 : w) : w / 2);
-        POINT loc = {x, y};
+        POINT loc = {static_cast<LONG>(x), static_cast<LONG>(y)};
         if (GetIsAutomap()) {
             loc = ScreenToAutomap(x, y);
         }
@@ -350,7 +350,7 @@ void ImageHook::Draw(void) {
     if (GetIsVisible() && GetX() != -1 && GetY() != -1 && GetImage() != NULL && image != NULL) {
         uint x = GetX(), y = GetY(), w = image->cells[0]->width;
         x += (alignment != Left ? (alignment != Right ? 0 : -1 * (w / 2)) : w / 2);
-        POINT loc = {x, y};
+        POINT loc = {static_cast<LONG>(x), static_cast<LONG>(y)};
         if (GetIsAutomap()) {
             loc = ScreenToAutomap(x, y);
         }
@@ -391,8 +391,8 @@ void LineHook::Draw(void) {
     Lock();
     if (GetIsVisible() && GetX() != -1 && GetY() != -1) {
         uint x = GetX(), y = GetY(), _x2 = GetX2(), _y2 = GetY2();
-        POINT loc = {x, y};
-        POINT sz = {_x2, _y2};
+        POINT loc = {static_cast<LONG>(x), static_cast<LONG>(y)};
+        POINT sz = {static_cast<LONG>(_x2), static_cast<LONG>(_y2)};
         if (GetIsAutomap()) {
             loc = ScreenToAutomap(x, y);
             sz = ScreenToAutomap(_x2, _y2);
@@ -413,8 +413,8 @@ void BoxHook::Draw(void) {
         } else if (alignment == Right) {
             x += x2 / 2;
         }
-        POINT loc = {x, y};
-        POINT sz = {x + x2, y + y2};
+        POINT loc = {static_cast<LONG>(x), static_cast<LONG>(y)};
+        POINT sz = {static_cast<LONG>(x + x2), static_cast<LONG>(y + y2)};
         if (GetIsAutomap()) {
             loc = ScreenToAutomap(x, y);
             sz = ScreenToAutomap(x + x2, y + y2);
@@ -442,7 +442,9 @@ void FrameHook::Draw(void) {
         } else if (alignment == Right) {
             x += x2 / 2;
         }
-        RECT rect = {x, y, x + x2, y + y2};
+        RECT rect = {static_cast<LONG>(x), static_cast<LONG>(y),
+                     static_cast<LONG>(x + x2),
+                     static_cast<LONG>(y + y2)};
         EnterCriticalSection(&Vars.cFrameHookSection);
         D2GFX_DrawFrame(&rect);
         LeaveCriticalSection(&Vars.cFrameHookSection);
